requirements.txt
Django==2.2.10
django-filter==2.4.0
djangorestframework==3.11.2
drf-yasg==1.17.0
gunicorn==20.1.0
python-dotenv==0.19.0
psycopg2-binary==2.8.6


